+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changes For IRIS Service
 unreleased
 ==========
 
+ - don't allow to get a deleted petition using the public API
  - hashtags for tweets: #petition #city-name
 
 2017/04/26 0.9.0